refactor useless import
move Constants class
